+++ clients/python-legacy/lakefs_client/api/staging_api.py
@@ -132,6 +132,7 @@ class StagingApi(object):
                     'branch',
                     'path',
                     'staging_metadata',
+                    'if_none_match',
                 ],
                 'required': [
                     'repository',
@@ -160,17 +161,21 @@ class StagingApi(object):
                         (str,),
                     'staging_metadata':
                         (StagingMetadata,),
+                    'if_none_match':
+                        (str,),
                 },
                 'attribute_map': {
                     'repository': 'repository',
                     'branch': 'branch',
                     'path': 'path',
+                    'if_none_match': 'If-None-Match',
                 },
                 'location_map': {
                     'repository': 'path',
                     'branch': 'path',
                     'path': 'query',
                     'staging_metadata': 'body',
+                    'if_none_match': 'header',
                 },
                 'collection_format_map': {
                 }
@@ -284,6 +289,7 @@ class StagingApi(object):
             staging_metadata (StagingMetadata):
 
         Keyword Args:
+            if_none_match (str): Set to \"*\" to atomically allow the upload only if the key has no object yet. Other values are not supported.. [optional]
             _return_http_data_only (bool): response data without head status
                 code and headers. Default is True.
             _preload_content (bool): if False, the urllib3.HTTPResponse object